Fix compilation on gcc/mingw
Gcc didn't properly support logf until recently despite it being in the standard, switch to log instead for broader compiler compatibility.

Also add an explicit cmath include, as MINGW couldn't find std::round in CI

+++ src/guiengine/widgets/kart_stats_widget.cpp
@@ -26,6 +26,7 @@
 #include "utils/translation.hpp"
 #include "io/file_manager.hpp"
 
+#include <cmath> // for log and round
 #include <IGUIEnvironment.h>
 #include <IGUIElement.h>
 #include <IGUIButton.h>
@@ -110,7 +111,7 @@ void KartStatsWidget::setValues(const KartProperties* props, HandicapLevel h)
     // We use a logarithm so that a given width corresponds to a given weight ratio
     // instead of a given absolute weight difference. This allows for a more meaningful
     // display, as the gameplay effects are proportional to the weight ratio.
-    float mass_log = std::logf(kpc->getMass());
+    float mass_log = std::log(kpc->getMass());
     float mass_display = (mass_log * 109.0f) - 545.0f;
     skill_pts += setSkillValues(SKILL_MASS, mass_display, "mass.png", "mass", _("Mass"));
